add ISSN validator + tests + documentation (#1166)
## International Standard Serial Number (ISSN) validator

"_An ISSN is an 8-digit code used to identify newspapers, journals,
magazines and periodicals of all kinds and on all media–print and
electronic._" --
[issn.org](https://www.issn.org/understanding-the-issn/what-is-an-issn/)

This PR adds a new `issn` validator, along with updating all translation
files and tests. This validator is very similar to that of the ISBN so I
believe all translations will be okay. The _valid_ ISSNs used in the
tests are examples taken from issn.org.

**Make sure that you've checked the boxes below before you submit PR:**
- [X] Tests exist or have been written that cover this particular
change.

@go-playground/validator-maintainers