fix: Revert min limit of e164 regex (#1516)
## Fixes Or Enhances

Although the E164 spec technically would allow for numbers with only one
digit after the country code, we used to expect atleast 7 digits, since
that is what is way more practical.

In fix for E164 to disallow leading 0 in country codes, we lowered said
limit to 1 digit. This broke things for peoples as numbers like `123456`
became valid.

Reverting back to the original limit of 7 digits, is practical, and
probably the expected behavior out of this validator.

Related #1515 

**Make sure that you've checked the boxes below before you submit PR:**
- [x] Tests exist or have been written that cover this particular
change.

@go-playground/validator-maintainers